backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
httpx==0.26.0
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
scikit-learn==1.4.0
numpy==1.26.3
openai==1.10.0
